Update 0 is 1

+++ singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int nodo1 = 0;
+	size_t nodo1 = 0;
 
 	while (h != NULL)
 	{
